+++ RELEASE_NOTES.md
@@ -25,6 +25,12 @@ RELEASE NOTES
 3.96.0 - November ??, 2025
 ==========================
 
+Prolog adapter and integration files
+------------------------------------
+
+* ADDED: Internal predicate, `'$lgt_time_stamp'/1`, for retrieving an opaque
+but comparable time stamp for the current time.
+
 IDEs, text editors, and syntax highlighters support
 ---------------------------------------------------